Added new rules to SftpConnection#listFiles()
Ensured that it is CDing to the new directory when listing files then CDing
back when finished. This ensures that the FULL PATH of the file is correct in relation
to the FTP user currently connected.

A bit nasty but current libs don't seem to have a built-in #getPath() method.